cloud_agnostic_storage/requirements_Dev.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosmtpd==1.4.2
amqp==5.0.6
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asgiref==3.4.1
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
atpublic==3.0.1
attrs==19.3.0
Automat==0.8.0
azure-core==1.24.2
azure-storage-blob==12.12.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
billiard==3.6.4.0
binaryornot==0.4.4
black==21.12b0
blinker==1.4
boto3==1.24.24
botocore==1.27.24
CacheControl==0.12.11
cachetools==4.2.4
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloud-init==22.1
cloudpickle==2.0.0
cmdkit==2.6.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
cvxopt==1.3.0
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.10
django-allauth==0.47.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-configurations==2.3.1
django-cors-headers==3.10.1
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
django-heroku==0.3.1
django-import-export==2.8.0
django-oauth-toolkit==1.7.0
django-storages==1.12.3
django-timezone-field==4.2.1
djangorestframework==3.12.4
dnspython==2.1.0
docutils==0.18.1
drf-psq==1.1.0
drf-yasg==1.20.0
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.3
fcm-django==1.0.11
firebase-admin==5.2.0
fonttools==4.31.2
frozenlist==1.3.0
geoip2==4.5.0
google-api-core==2.7.2
google-api-python-client==2.46.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
gym==0.23.1
gym-notices==0.0.6
h11==0.13.0
httpie==1.0.3
httplib2==0.20.4
hyperlink==19.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==16.10.1
inflection==0.5.1
ipython==7.30.1
isodate==0.6.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==1.0.1
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jwcrypto==1.0
keyring==18.0.1
kiwisolver==1.4.2
kombu==5.2.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
maxminddb==2.2.0
more-itertools==4.2.0
msgpack==1.0.3
msrest==0.7.1
multidict==6.0.2
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.1
outcome==1.1.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
platformdirs==2.4.0
poyo==0.5.0
progressbar33==2.4
prompt-toolkit==3.0.24
proto-plus==1.20.3
protobuf==3.20.1
psycopg2==2.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.3.0
pylibmc==1.6.1
pymacaroons==0.13.0
pymemcache==3.5.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-slugify==6.1.1
python3-openid==3.2.0
pytz==2021.3
pywatchman==1.4.1
PyYAML==6.0
redis==4.2.2
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.8.0
SecretStorage==2.3.1
selenium==4.1.3
sentry-sdk==1.5.12
service-identity==18.1.0
simplejson==3.16.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
sos==4.3
sqlparse==0.4.2
ssh-import-id==5.10
stack-data==0.2.0
systemd-python==234
tablib==3.2.1
tblib==1.7.0
terminaltables==3.1.10
text-unidecode==1.3
threadpoolctl==3.1.0
tomli==1.2.2
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
Twisted==18.9.0
typing-extensions==4.0.1
tzdata==2022.1
ubuntu-advantage-tools==27.8
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
websocket-client==1.2.3
whitenoise==6.0.0
wrapt==1.14.0
wsproto==1.1.0
yarl==1.7.2
zipp==1.0.0
zope.interface==4.7.1
